консольный файловый менеджер готов
